+++ packages/terra-data-grid/src/index.js
@@ -1,4 +1,5 @@
 import WorklistDataGrid from './WorklistDataGrid';
+import DataGrid from './DataGrid';
 
 // eslint-disable-next-line import/prefer-default-export
-export { WorklistDataGrid };
+export { WorklistDataGrid, DataGrid };